src/DisplayServer/requirements.txt
beautifultable==1.1.0
bleach==6.1.0
blinker==1.7.0
chardet==5.2.0
click==8.1.7
cssselect2==0.7.0
filelock==3.13.1
Flask-Cors==4.0.0
flask==3.0.2
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
pillow==10.2.0
pip==23.2.1
pyinkscape==0.1a3
pysondb==1.6.7
PyYAML==6.0.1
reportlab==4.1.0
setuptools==68.2.0
six==1.16.0
svg2rlg==0.3
svglib==1.5.1
tinycss2==1.2.1
typer==0.9.0
typing-extensions==4.10.0
waitress==3.0.0
Wand==0.6.13
wcwidth==0.2.13
webencodings==0.5.1
werkzeug==3.0.1
wheel==0.41.2
zipp==3.17.0
rlPyCairo==0.3.0
cairosvg==2.7.1
svgpathtools==1.6.1
requests==2.31.0
python-magic==0.4.27
numpy==1.24.4
matplotlib==3.7.5